requirements.txt
cryptography>=41.0.0
pyyaml>=6.0
click>=8.1.0
flask>=3.0.0
werkzeug>=3.0.0
qrcode[pil]>=7.4.2
pillow>=10.0.0

